Logistics 8b.3: order suggestion engine
Ports routes/performance.js's computeOrderSuggestions/computeVendorOrderBuild/
enforceMaterialQty/validateVendorCombinedQty/buildAcceptPayload and the full
/order-suggestions/* route set (list/build/preview/accept/accept-batch/manual/
manual-bulk/tracked/edit/patch-po-item/delete/assign-schedule-agreement) onto
PurchaseOrderSuggestionHelper. Every pure business function (lot-size
enforcement, vendor combined-quantity validation, accept-payload date math,
the Urgency state machine, vendor grouping) is unit-tested in isolation.

Fixes a real fidelity gap found while wiring this slice: ForecastMathHelper.
DemandOverDays was typed `int days` but the caller's coverage-days figure is
genuinely fractional whenever a vendor's lead time is (Node adds fractional
days with no truncation) — widened to `decimal days`.

create-po/regenerate-pdf (real elevated SAP PO creation) remain deferred to
8b.7; Inbound Shipment tracking is 8b.4.

+++ dotnet/NormantonNexus/Helpers/Logistics/ForecastMathHelper.cs
@@ -212,9 +212,9 @@ internal static class ForecastMathHelper
     /// it does to the graph — used for the order-suggestion sizing calculation, which
     /// needs a demand total over an arbitrary day-count rather than the bucketed series.
     /// </summary>
-    internal static decimal DemandOverDays(IReadOnlyList<decimal> predictedMonthly, DateTime from, int days, IReadOnlyList<DemandAdjustmentWindow>? adjustments = null, Func<DateTime, decimal>? dailyUsageFnOverride = null)
+    internal static decimal DemandOverDays(IReadOnlyList<decimal> predictedMonthly, DateTime from, decimal days, IReadOnlyList<DemandAdjustmentWindow>? adjustments = null, Func<DateTime, decimal>? dailyUsageFnOverride = null)
     {
-        var rangeEnd = from.AddDays(days);
+        var rangeEnd = from.AddDays((double)days);
         var dailyUsage = dailyUsageFnOverride ?? MakeDailyUsageFn(predictedMonthly, from, adjustments ?? []);
         var total = 0m;
         for (var day = from; day < rangeEnd; day = day.AddDays(1))
